Added self signed cert bootstrap
In order for an Nginx server to be able to start when configured with SSL there needs to be a certificate in place. This adds a step to generate a selfsigned certificate of the proper name so that the Nginx process can start properly in order for the letsencrypt process to run.